nlp/pyproject.toml
[project]
name = "stratos-nlp"
version = "0.1.0"
description = "STRATOS NLP & LLM Stack — sentiment, embeddings, RAG"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "transformers>=4.38",
    "sentence-transformers>=2.5",
    "spacy>=3.7",
    "langchain>=0.1",
    "pgvector>=0.2",
    "structlog>=24.1",
    "langchain-huggingface>=0.0.1",
    "langchain-community>=0.0.1",
    "faiss-cpu>=1.7",
]
[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3", "mypy>=1.8"]
[tool.ruff]
target-version = "py312"
line-length = 100
